requirements.txt
streamlit
langchain
langchain-groq
sqlalchemy
openai-whisper
gTTS
pyaudio
sounddevice
soundfile
python-dotenv
pandas